Added a Recipe Interface
